requirements.txt
pyyaml
sqlalchemy
psycopg2
requests
beautifulsoup4
